% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{BEDMatrix-package}
\title{BEDMatrix: A Wrapper for Binary PED Files}
\description{
The BEDMatrix package provides a wrapper around binary PED files (so-called
BED files) that behaves just like a regular R matrix, but retrieves
genotypes on demand without loading the entire BED file into memory. The
goal is to support huge PED files, and to save time on initially reading in
PED files into the R environment.
}

