% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDMatrix.R
\docType{methods}
\name{show,BEDMatrix-method}
\alias{show,BEDMatrix-method}
\title{Show a BEDMatrix Object.}
\usage{
\S4method{show}{BEDMatrix}(object)
}
\arguments{
\item{object}{A \linkS4class{BEDMatrix} object.}
}
\description{
Display the object, by printing, plotting or whatever suits its class.
}
