% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{BEDMatrix-package}
\title{BEDMatrix: A Wrapper for Binary PED Files}
\description{
The BEDMatrix package provides a wrapper around binary PED (also known as
BED) files, one of the genotype/phenotype file formats of
\href{https://www.cog-genomics.org/plink2}{PLINK}, the whole genome association
analysis toolset. \linkS4class{BEDMatrix} objects are created by simply providing
the path to a BED file and once created, they behave similarly to regular
matrices with the advantage that genotypes are retrieved on demand without
loading the entire file into memory. This allows handling of very large
files with limited use of memory. Technically, a \linkS4class{BEDMatrix} object is
a memory-mapped matrix backed by a binary PED file.
}
